Refactor: Move common layout elements to CommonWrapper
This commit introduces a new component, `CommonWrapper`, to encapsulate the common layout
elements such as `Navigation`, `QuickMenu`, `PlayerCard`, and `Footer`. This reduces
duplication in layout files and centralizes layout management.

Additionally, the `QueryClient` is now instantiated in the `+layout.svelte` file and passed to
the `QueryClientProvider`. This ensures that the `QueryClient` is only created once and is
available to all components in the application.